requirements.txt
aniso8601==2.0.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
cffi==1.11.4
click==6.7
constantly==15.1.0
cssselect==1.0.3
coverage==4.5.1
dateparser==0.7.0
fake-useragent==0.1.10
hyperlink==17.3.1
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.2
numpy==1.14.1
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
queuelib==1.4.2
regex==2018.2.21
Scrapy==1.5.0
scrapy-fake-useragent==1.1.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
json-lines
matplotlib
python-igraph
cairocffi
neovim-gui
flask
pytest
Flask-Testing